+++ cookbooks/ruby_build/metadata.rb
@@ -2,8 +2,8 @@ name              "ruby_build"
 maintainer        "Fletcher Nichol"
 maintainer_email  "fnichol@nichol.ca"
 license           "Apache 2.0"
-description       "Manages the ruby-build framework and its installed rubies. A LWRP is also defined."
-long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+descriptioin       "Manages the ruby-build framework and its installed rubies. A LWRP is also defined."
+long_descriptioin  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version           "0.8.0"
 
 supports "ubuntu"