+++ cookbooks/yum-mysql-community/attributes/mysql57-community.rb
@@ -1,6 +1,6 @@
 default['yum']['mysql57-community']['repositoryid'] = 'mysql57-community'
 default['yum']['mysql57-community']['gpgkey'] = 'https://raw.githubusercontent.com/chef-cookbooks/yum-mysql-community/master/files/default/mysql_pubkey.asc'
-default['yum']['mysql57-community']['description'] = 'MySQL 5.7 Community Server'
+default['yum']['mysql57-community']['descriptioin'] = 'MySQL 5.7 Community Server'
 default['yum']['mysql57-community']['failovermethod'] = 'priority'
 default['yum']['mysql57-community']['gpgcheck'] = true
 default['yum']['mysql57-community']['enabled'] = true